# docs/reference/openapi.smartout.internal.yaml
openapi: 3.1.0
info:
  title: Smartout API (internal)
  version: 0.1.0-draft-internal
  summary: Starter OpenAPI contract for Smartout internal and target external APIs.
  description: >
    This starter spec captures currently implemented endpoints and planned `/v1`
    endpoint groups.

    It is intentionally contract-first and roadmap-aware.
servers:
  - url: https://app.smartout.ai
    description: Web app host (Next.js API routes)
  - url: https://smartout.ai
    description: Landing host (Next.js API routes)
  - url: https://{projectRef}.supabase.co/functions/v1
    description: Supabase Edge Functions
    variables:
      projectRef:
        default: your-project-ref
  - url: http://localhost:8000
    description: Internal scrapling service (local/internal)
...
paths:
  /api/wizard/start:
    post:
      tags:
        - Voice
      operationId: startVoiceMission
      summary: Start Ultravox voice mission
      description: Used by both web and landing apps; mission default differs by caller.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WizardStartRequest"
      responses:
        "200":
          description: Voice call started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WizardStartResponse"
        "502":
          $ref: "#/components/responses/BadGatewayError"
        "503":
          $ref: "#/components/responses/ServiceUnavailableError"
      x-smartout-surface: internal
      x-roadmap-status: implemented
...
  /api/docs-agent:
    post:
      tags:
        - Onboarding
      operationId: askDocsAgent
      summary: Ask docs assistant (landing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 4000
                history:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum:
                          - user
                          - assistant
                      content:
                        type: string
      responses:
        "200":
          description: Answer with sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        href:
                          type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-smartout-surface: internal
      x-roadmap-status: implemented
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
    DocuSealSignature:
      type: apiKey
      in: header
      name: x-docuseal-signature
...
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGatewayError:
      description: Upstream gateway failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailableError:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
    WizardStartRequest:
      type: object
      properties:
        mission_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
    WizardStartResponse:
      type: object
      properties:
        joinUrl:
          type: string
        callId:
          type: string
        mission:
          type: string
      required:
        - joinUrl
        - callId
        - mission
...